Extract ManifestUtil for reuse
