+++ src/main/groovy/com/bugsnag/android/gradle/BugsnagManifestTask.groovy
@@ -15,10 +15,6 @@ import org.gradle.api.tasks.TaskAction
  */
 class BugsnagManifestTask extends BugsnagVariantOutputTask {
 
-    private static final String TAG_META_DATA = 'meta-data'
-    private static final String NS_URI_ANDROID = "http://schemas.android.com/apk/res/android"
-    private static final String NS_PREFIX_ANDROID = "android"
-
     BugsnagManifestTask() {
         super()
         this.description = "Adds a unique build UUID to AndroidManifest to link proguard mappings to crash reports"
@@ -36,46 +32,7 @@ class BugsnagManifestTask extends BugsnagVariantOutputTask {
                 continue
             }
 
-            project.logger.debug("Updating manifest with build UUID: " + manifestPath)
-
-            // Parse the AndroidManifest.xml
-            Namespace ns = new Namespace(NS_URI_ANDROID, NS_PREFIX_ANDROID)
-            Node xml = new XmlParser().parse(manifestPath)
-
-            def application = xml.application[0]
-            if (application) {
-                def metaDataTags = application[TAG_META_DATA]
-
-                // if BUILD_UUID is already present don't override what user has specified
-                def tags = metaDataTags.findAll {
-                    (it.attributes()[ns.name] == BugsnagPlugin.BUILD_UUID_TAG)
-                }
-                if (tags.size() > 0) {
-                    return
-                }
-
-                // Add the new BUILD_UUID_TAG element
-                application.appendNode(TAG_META_DATA, [(ns.name):BugsnagPlugin.BUILD_UUID_TAG, (ns.value):buildUUID])
-
-                // Write the manifest file
-                FileWriter writer = null
-
-                try {
-                    writer = new FileWriter(manifestPath)
-                    XmlNodePrinter printer = new XmlNodePrinter(new PrintWriter(writer))
-                    printer.preserveWhitespace = true
-                    printer.print(xml)
-                } catch (IOException e) {
-                    project.logger.warn("Failed to update manifest with Bugsnag metadata", e)
-                } finally {
-                    if (writer != null) {
-                        writer.close()
-                    }
-                }
-            } else {
-                project.logger.error("Bugsnag detected invalid manifest with no " +
-                    "application element so did not write Build UUID")
-            }
+            ManifestUtil.patchManifest(manifestPath, buildUUID, logger)
         }
     }
